+++ bobtemplates/cs/bobregistry.py
@@ -15,3 +15,11 @@ def cs_plone_buildout():
     reg.template = 'bobtemplates.cs:cs_plone_buildout'
     reg.plonecli_alias = 'cs_plone_buildout'
     return reg
+
+
+def cs_migration():
+    reg = RegEntry()
+    reg.template = 'bobtemplates.cs:cs_migration'
+    reg.plonecli_alias = 'cs_migration'
+    reg.depend_on = 'plone_addon'
+    return reg